Refactor name of utilities
